feat(cli): add new plugins for lazyvim
Add the `lazyvim.plugins.extras.lang.json` and `lazyvim.plugins.extras.ui.mini-animate` plugins to the lazyvim configuration in `nvim.nix`. This allows for JSON language support and a mini animation UI feature.

feat(homebrew): add battle-net cask

Add the `battle-net` cask to the list of Homebrew casks in the `homebrew.nix` file. This allows for installation of the Blizzard launcher.